java script test

+++ js/res.js
@@ -7,13 +7,66 @@ if(localStorage.getItem("ChristamsGifts")){
   var giftSourse = document.getElementById('SearchResult');
   
   var forget = localStorage.getItem('ChristamsGifts');
-  var farhanx = JSON.parse(forget);
+  var yy = JSON.parse(forget);
 
-  console.log(farhanx);
-  for(var i = 0 ; i < farhanx.length;i++){
-    var fathi  =document.createElement('img');
-    fathi.src=farhanx[i];
-    try1.appendChild(fathi); 
+  console.log(yy);
+  for(var i = 0 ; i < yy.length;i++){
+    var xx  =document.createElement('img');
+    xx.src=yy[i];
+    try1.appendChild(xx); 
+  } 
+  localStorage.clear();
+}
+
+if(localStorage.getItem("EidGifts")){
+  
+ 
+  var giftSourse = document.getElementById('SearchResult');
+  
+  var forget = localStorage.getItem('EidGifts');
+  var yy = JSON.parse(forget);
+
+  console.log(yy);
+  for(var i = 0 ; i < yy.length;i++){
+    var xx  =document.createElement('img');
+    xx.src=yy[i];
+    try1.appendChild(xx); 
+  } 
+  localStorage.clear();
+}
+
+
+if(localStorage.getItem("GetWellGifts")){
+  
+ 
+  var giftSourse = document.getElementById('SearchResult');
+  
+  var forget = localStorage.getItem('GetWellGifts');
+  var yy = JSON.parse(forget);
+
+  console.log(yy);
+  for(var i = 0 ; i < yy.length;i++){
+    var xx  =document.createElement('img');
+    xx.src=yy[i];
+    try1.appendChild(xx); 
+  } 
+  localStorage.clear();
+}
+
+
+if(localStorage.getItem("BirthdayGifts")){
+  
+ 
+  var giftSourse = document.getElementById('SearchResult');
+  
+  var forget = localStorage.getItem('BirthdayGifts');
+  var yy = JSON.parse(forget);
+
+  console.log(yy);
+  for(var i = 0 ; i < yy.length;i++){
+    var xx  =document.createElement('img');
+    xx.src=yy[i];
+    try1.appendChild(xx); 
   } 
   localStorage.clear();
 }
